# crimson/os/seastore/object_data_handler: make sure continuous object data blocks are loaded with a single io request