… src/crimson/os/seastore/segment_manager/ephemeral.cc
… src/crimson/os/seastore/segment_manager/zbd.cc
… src/crimson/os/seastore/segment_manager/zbd.h
… src/crimson/os/seastore/transaction_manager.h
… src/test/crimson/seastore/test_object_data_handler.cc