@athanatos 2025-09-22
This PR looks correct, but it's a *lot* of additional code.  Can we at least reimplement the existing interfaces in terms of the new aggregate ones?

Also, do you have an idea of how much this helps?  We've been trying to let extents be large -- I'd have guessed that would be sufficient to ensure that devices reads were reasonably large?