@xxhdx1985126 2025-09-24
> This PR looks correct, but it's a _lot_ of additional code. Can we at least reimplement the existing interfaces in terms of the new aggregate ones?

Do you mean that instead of adding new interfaces, we adjust the current interfaces to support aggregate extents reads?

> 
> Also, do you have an idea of how much this helps? We've been trying to let extents be large -- I'd have guessed that would be sufficient to ensure that devices reads were reasonably large?

This issue came up in our online cluster, in which we use HDD as a cold tier. The large read IOs on the cold tier was very slow, with the latency 900+ ms. After analysis, we found that the reasons are as follows:
1. as we needed to do full extent integrity check, we set seastore_max_data_allocation_size to 32KB limiting the read amplification, so large extents are cut into 32KB pieces;
2. when reading large blocks of data from the cold tier, one large read were transformed into a series of 32KB reads, which were done sequentially.

After the we implemented the machinery in this PR, the latency was reduced to less than 50ms.

Here is our test on SSD machines, which shows even with the hot tier, this PR can provide about 20% large block read performance improvement:
**Main Branch**
```
read-128k: (g=0): rw=read, bs=(R) 4096KiB-4096KiB, (W) 4096KiB-4096KiB, (T) 4096KiB-4096KiB, ioengine=rbd, iodepth=1
fio-3.33-61-g4483
Starting 1 process
Jobs: 1 (f=0): [f(1)][100.0%][r=80.1MiB/s][r=20 IOPS][eta 00m:00s]
read-128k: (groupid=0, jobs=1): err= 0: pid=233295: Wed Sep 24 10:40:31 2025
  read: IOPS=30, BW=124MiB/s (130MB/s)(10.0GiB/82856msec)
    slat (usec): min=6, max=110, avg= 9.58, stdev= 3.43
    clat (usec): min=20585, max=76303, avg=32353.69, stdev=3435.32
     lat (usec): min=20593, max=76311, avg=32363.27, stdev=3435.36
    clat percentiles (usec):
     |  1.00th=[27919],  5.00th=[29492], 10.00th=[30016], 20.00th=[30540],
     | 30.00th=[31065], 40.00th=[31589], 50.00th=[32113], 60.00th=[32637],
     | 70.00th=[33162], 80.00th=[33817], 90.00th=[34341], 95.00th=[34866],
     | 99.00th=[45351], 99.50th=[58459], 99.90th=[72877], 99.95th=[73925],
     | 99.99th=[76022]
   bw (  KiB/s): min=73728, max=139264, per=100.00%, avg=126752.58, stdev=8220.11, samples=165
   iops        : min=   18, max=   34, avg=30.95, stdev= 2.01, samples=165
  lat (msec)   : 50=99.26%, 100=0.74%
  cpu          : usr=0.06%, sys=0.03%, ctx=2562, majf=0, minf=189
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=2560,0,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
   READ: bw=124MiB/s (130MB/s), 124MiB/s-124MiB/s (130MB/s-130MB/s), io=10.0GiB (10.7GB), run=82856-82856msec

Disk stats (read/write):
  nvme1n1: ios=0/1008, merge=0/1238, ticks=0/119, in_queue=119, util=0.67%
```

**This PR**
```
read-128k: (g=0): rw=read, bs=(R) 4096KiB-4096KiB, (W) 4096KiB-4096KiB, (T) 4096KiB-4096KiB, ioengine=rbd, iodepth=1
fio-3.37-124-gbcd4
Starting 1 process
Jobs: 1 (f=1): [R(1)][100.0%][r=84.0MiB/s][r=21 IOPS][eta 00m:00s]
read-128k: (groupid=0, jobs=1): err= 0: pid=2862965: Wed Sep 24 10:37:32 2025
  read: IOPS=37, BW=149MiB/s (156MB/s)(10.0GiB/68702msec)
    slat (nsec): min=4450, max=50520, avg=5385.18, stdev=1623.30
    clat (usec): min=22340, max=78659, avg=26830.15, stdev=3553.71
     lat (usec): min=22345, max=78664, avg=26835.54, stdev=3553.75
    clat percentiles (usec):
     |  1.00th=[23200],  5.00th=[23725], 10.00th=[24249], 20.00th=[25035],
     | 30.00th=[25822], 40.00th=[26346], 50.00th=[26870], 60.00th=[27132],
     | 70.00th=[27657], 80.00th=[27919], 90.00th=[28443], 95.00th=[28705],
     | 99.00th=[36439], 99.50th=[54264], 99.90th=[74974], 99.95th=[77071],
     | 99.99th=[79168]
   bw (  KiB/s): min=81920, max=163840, per=100.00%, avg=152777.81, stdev=11308.98, samples=137
   iops        : min=   20, max=   40, avg=37.30, stdev= 2.76, samples=137
  lat (msec)   : 50=99.41%, 100=0.59%
  cpu          : usr=0.05%, sys=0.02%, ctx=2562, majf=0, minf=81
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=2560,0,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
   READ: bw=149MiB/s (156MB/s), 149MiB/s-149MiB/s (156MB/s-156MB/s), io=10.0GiB (10.7GB), run=68702-68702msec

Disk stats (read/write):
  nvme0n1: ios=0/817, sectors=0/24000, merge=0/1145, ticks=0/81, in_queue=80, util=0.59%
```