@xxhdx1985126 2025-09-24
Sorry, I tested the wrong branch, it was actually 400% large block read perf improvement. I've just updated https://github.com/ceph/ceph/pull/65425#issuecomment-3326236943