@Matan-B 2025-10-15
I've looked at 
https://tracker.ceph.com/issues/73354:
```
void crimson::os::seastore::CachedExtent::set_io_wait(extent_state_t)', ceph_assert(!io_wait)
```
Seems also related to me - I'll close the tracker since the PR is not yet merged.