@Matan-B 2025-12-04
```
/ceph/src/crimson/os/seastore/random_block_manager/nvme_block_device.cc:171:3: error: use of undeclared identifier 'logger'; did you mean 'seastar::logger'?
```

@xxhdx1985126, is this ready for retesting?