@xxhdx1985126 2026-04-02
> ```
> [==========] 128 tests from 1 test suite ran. (346675 ms total)
> [  PASSED  ] 124 tests.
> [  FAILED  ] 4 tests, listed below:
> [  FAILED  ] object_data_handler_test/object_data_handler_test_t.aggregate_read/0, where GetParam() = (0x6015a41e6680 pointing to "segmented", 1-byte object <00>)
> [  FAILED  ] object_data_handler_test/object_data_handler_test_t.aggregate_read/1, where GetParam() = (0x6015a41e6680 pointing to "segmented", 1-byte object <01>)
> [  FAILED  ] object_data_handler_test/object_data_handler_test_t.aggregate_read/2, where GetParam() = (0x6015a41e66a0 pointing to "circularbounded", 1-byte object <00>)
> [  FAILED  ] object_data_handler_test/object_data_handler_test_t.aggregate_read/3, where GetParam() = (0x6015a41e66a0 pointing to "circularbounded", 1-byte object <01>)
> ```
> 
> @xxhdx1985126

Just fixed, this was caused by uninitialized `known_contents` in the unittest.